requirements.txt
einops==0.7.0
librosa==0.9.2
matplotlib==3.7.0
numpy==1.23.5
onnxruntime_gpu==1.9.0
pandas==1.5.3
pesq==0.0.4
PyYAML==6.0
Requests==2.31.0
scipy==1.12.0
SoundFile==0.10.3.post1
tensorboardX==2.6
torch==2.0.0
torchaudio==2.0.1
tqdm==4.64.1
